tasks.ceph_deploy: Work around admin keyring bug
This is a short term work around for bug:

    http://tracker.ceph.com/issues/17849

As requested by Sage on the ceph-devel mailing list in thread:

    "Understanding encrypted OSD's and cephx"

The work around adds the admin key to all nodes with encrypted OSD's by
running the command:

    ceph-deploy admin $NODE

So the admin keyring is present.

This work around should be fixed by making a new "one step set up" command,
as detailed by Sage.

Signed-off-by: Owen Synge <osynge@suse.com>